requirements.txt
pywin32
psutil~=5.9.8
pyinstaller
pystray~=0.19.5
pillow~=10.3.0
win11toast